updated farm design
This manual should serve as a comprehensive guide for both industry professionals and individuals interested in exploring the seaweed farming space with the potential of starting their own small-scale project.

Please keep in mind that it is still WIP and multiple updates and edits will be made in the future.